Extract damage handling to EntityDamageHandler
