+++ src/main/java/me/ryanhamshire/GriefPrevention/PlayerEventHandler.java
@@ -1356,7 +1356,7 @@ class PlayerEventHandler implements Listener
         if (itemInHand.getType() == Material.NAME_TAG)
         {
             EntityDamageByEntityEvent damageEvent = new EntityDamageByEntityEvent(player, entity, EntityDamageEvent.DamageCause.CUSTOM, 0);
-            instance.entityEventHandler.onEntityDamage(damageEvent);
+            instance.entityDamageHandler.onEntityDamage(damageEvent);
             if (damageEvent.isCancelled())
             {
                 event.setCancelled(true);
